update paddle/tensor, test=develop
