+++ python/paddle/fluid/tests/unittests/test_addcmul.py
@@ -37,7 +37,7 @@ class TestAddcmulLayer(unittest.TestCase):
             tensor1 = fluid.data(name="tensor1", dtype=self._dtype, shape=[100])
             tensor2 = fluid.data(
                 name="tensor2", dtype=self._dtype, shape=[3, 100])
-            out = paddle.addcmul(input, tensor1, tensor2, value)
+            out = fluid.layers.addcmul(input, tensor1, tensor2, value)
 
         exe = fluid.Executor(self._place)
         return exe.run(feed={
@@ -53,7 +53,7 @@ class TestAddcmulLayer(unittest.TestCase):
             input = fluid.dygraph.to_variable(self.input)
             tensor1 = fluid.dygraph.to_variable(self.tensor1)
             tensor2 = fluid.dygraph.to_variable(self.tensor2)
-            out = paddle.addcmul(input, tensor1, tensor2, value)
+            out = fluid.layers.addcmul(input, tensor1, tensor2, value)
             return out.numpy()
 
     def numpy(self, value=1.0):
@@ -85,7 +85,7 @@ class TestAddcmul(unittest.TestCase):
             tensor1 = fluid.data(name='t1', shape=data_shape, dtype='float32')
             tensor2 = fluid.data(name='t2', shape=data_shape, dtype='float32')
 
-            out = paddle.addcmul(input, tensor1, tensor2)
+            out = fluid.layers.addcmul(input, tensor1, tensor2)
             self.assertEqual(out.shape, input.shape)
 
     def test_addcmul_with_broadcast0(self):
@@ -95,7 +95,7 @@ class TestAddcmul(unittest.TestCase):
             tensor1 = fluid.data(name='t1', shape=[3, 100], dtype='float32')
             tensor2 = fluid.data(name='t2', shape=[100], dtype='float32')
 
-            out = paddle.addcmul(input, tensor1, tensor2)
+            out = fluid.layers.addcmul(input, tensor1, tensor2)
             self.assertEqual(out.shape, input.shape)
 
     def test_addcmul_with_broadcast1(self):
@@ -105,7 +105,7 @@ class TestAddcmul(unittest.TestCase):
             tensor1 = fluid.data(name='t1', shape=[100], dtype='float32')
             tensor2 = fluid.data(name='t2', shape=[4, 100], dtype='float32')
 
-            out = paddle.addcmul(input, tensor1, tensor2)
+            out = fluid.layers.addcmul(input, tensor1, tensor2)
             self.assertEqual(out.shape, input.shape)
 
     def test_addcmul_with_broadcast2(self):
@@ -115,7 +115,7 @@ class TestAddcmul(unittest.TestCase):
             tensor1 = fluid.data(name='t1', shape=[100], dtype='float32')
             tensor2 = fluid.data(name='t2', shape=[100], dtype='float32')
 
-            out = paddle.addcmul(input, tensor1, tensor2)
+            out = fluid.layers.addcmul(input, tensor1, tensor2)
             self.assertEqual(out.shape, input.shape)
 
     def test_addcmul_has_out(self):
@@ -126,7 +126,7 @@ class TestAddcmul(unittest.TestCase):
             tensor2 = fluid.data(name='t2', shape=[100], dtype='float32')
             out = fluid.data(name='out', shape=[4, 100], dtype='float32')
 
-            out = paddle.addcmul(input, tensor1, tensor2, out=out)
+            out = fluid.layers.addcmul(input, tensor1, tensor2, out=out)
             self.assertEqual(out.shape, input.shape)
 
 
@@ -140,7 +140,7 @@ class InvalidInputTest(unittest.TestCase):
                     name='tensor1', shape=[20, 20], dtype='float32')
                 tensor2 = fluid.data(
                     name='tensor2', shape=[20, 20], dtype='float32')
-                out = paddle.addcmul(input, tensor1, tensor2)
+                out = fluid.layers.addcmul(input, tensor1, tensor2)
 
         self.assertRaises(TypeError, test_invalid_input)
 
@@ -152,7 +152,7 @@ class InvalidInputTest(unittest.TestCase):
                 tensor1 = [20, 20]
                 tensor2 = fluid.data(
                     name='tensor2', shape=[20, 20], dtype='float32')
-                out = paddle.addcmul(input, tensor1, tensor2)
+                out = fluid.layers.addcmul(input, tensor1, tensor2)
 
         self.assertRaises(TypeError, test_invalid_tensor1)
 
@@ -164,7 +164,7 @@ class InvalidInputTest(unittest.TestCase):
                 tensor1 = fluid.data(
                     name='tensor1', shape=[20, 20], dtype='float32')
                 tensor2 = [20, 20]
-                out = paddle.addcmul(input, tensor1, tensor2)
+                out = fluid.layers.addcmul(input, tensor1, tensor2)
 
         self.assertRaises(TypeError, test_invalid_tensor2)
 
@@ -177,7 +177,7 @@ class InvalidInputTest(unittest.TestCase):
                     name='tensor1', shape=[20, 20], dtype='float32')
                 tensor2 = fluid.data(
                     name='tensor2', shape=[20, 20], dtype='float32')
-                out = paddle.addcmul(input, tensor1, tensor2, value=1)
+                out = fluid.layers.addcmul(input, tensor1, tensor2, value=1)
 
         self.assertRaises(TypeError, test_invalid_value_int)
 
@@ -189,7 +189,7 @@ class InvalidInputTest(unittest.TestCase):
                     name='tensor1', shape=[20, 20], dtype='int32')
                 tensor2 = fluid.data(
                     name='tensor2', shape=[20, 20], dtype='int32')
-                out = paddle.addcmul(input, tensor1, tensor2, value=1.0)
+                out = fluid.layers.addcmul(input, tensor1, tensor2, value=1.0)
 
         self.assertRaises(TypeError, test_invalid_value_float)
