+++ python/paddle/fluid/tests/unittests/test_bmm_op.py
@@ -46,7 +46,7 @@ class API_TestBmm(unittest.TestCase):
                 'data1', shape=[-1, 3, 4], dtype='float64')
             data2 = fluid.layers.data(
                 'data2', shape=[-1, 4, 5], dtype='float64')
-            result_bmm = paddle.bmm(data1, data2)
+            result_bmm = fluid.layers.bmm(data1, data2)
             place = fluid.CPUPlace()
             exe = fluid.Executor(place)
             input1 = np.random.random([10, 3, 4]).astype('float64')
@@ -67,7 +67,7 @@ class API_TestDygraphBmm(unittest.TestCase):
         with fluid.dygraph.guard():
             x = fluid.dygraph.to_variable(input1)
             y = fluid.dygraph.to_variable(input2)
-            out = paddle.bmm(x, y)
+            out = fluid.layers.bmm(x, y)
             out_np = out.numpy()
         expected_result = np.matmul(input1, input2)
         self.assertTrue(np.allclose(expected_result, out_np))
