+++ python/paddle/fluid/tests/unittests/test_roll_op.py
@@ -60,7 +60,7 @@ class TestRollAPI(unittest.TestCase):
         self.data_x = np.array(
             [[1.0, 2.0, 3.0], [4.0, 5.0, 6.0], [7.0, 8.0, 9.0]])
 
-    def test_index_select_api(self):
+    def test_roll_api(self):
         self.input_data()
 
         # case 1:
