+++ python/paddle/tensor/__init__.py
@@ -71,59 +71,7 @@ from .random import randn  #DEFINE_ALIAS
 # from .random import rand   #DEFINE_ALIAS
 from .random import randint  #DEFINE_ALIAS
 from .random import randperm
-# from .math import abs   #DEFINE_ALIAS
-# from .math import acos   #DEFINE_ALIAS
-# from .math import asin   #DEFINE_ALIAS
-from .math import atan  #DEFINE_ALIAS
-# from .math import ceil   #DEFINE_ALIAS
-# from .math import cos   #DEFINE_ALIAS
-# from .math import cumsum   #DEFINE_ALIAS
-# from .math import elementwise_add   #DEFINE_ALIAS
-# from .math import elementwise_div   #DEFINE_ALIAS
-# from .math import elementwise_floordiv   #DEFINE_ALIAS
-# from .math import elementwise_max   #DEFINE_ALIAS
-# from .math import elementwise_min   #DEFINE_ALIAS
-# from .math import elementwise_mod   #DEFINE_ALIAS
-# from .math import elementwise_mul   #DEFINE_ALIAS
-# from .math import elementwise_pow   #DEFINE_ALIAS
-# from .math import elementwise_sub   #DEFINE_ALIAS
-# from .math import exp   #DEFINE_ALIAS
-# from .math import floor   #DEFINE_ALIAS
-# from .math import increment   #DEFINE_ALIAS
-# from .math import log   #DEFINE_ALIAS
-from .math import mul  #DEFINE_ALIAS
-# from .math import multiplex   #DEFINE_ALIAS
-from .math import pow  #DEFINE_ALIAS
-# from .math import reciprocal   #DEFINE_ALIAS
-# from .math import reduce_max   #DEFINE_ALIAS
-# from .math import reduce_min   #DEFINE_ALIAS
-# from .math import reduce_prod   #DEFINE_ALIAS
-# from .math import reduce_sum   #DEFINE_ALIAS
-# from .math import round   #DEFINE_ALIAS
-# from .math import rsqrt   #DEFINE_ALIAS
-# from .math import scale   #DEFINE_ALIAS
-# from .math import sign   #DEFINE_ALIAS
-from .math import sin  #DEFINE_ALIAS
-from .math import sqrt  #DEFINE_ALIAS
-# from .math import square   #DEFINE_ALIAS
-# from .math import stanh   #DEFINE_ALIAS
-from .math import sum  #DEFINE_ALIAS
-# from .math import sums   #DEFINE_ALIAS
-from .math import tanh  #DEFINE_ALIAS
-from .math import elementwise_sum  #DEFINE_ALIAS
-from .math import max  #DEFINE_ALIAS
-from .math import min  #DEFINE_ALIAS
-from .math import mm  #DEFINE_ALIAS
-from .math import div  #DEFINE_ALIAS
-from .math import add  #DEFINE_ALIAS
-# from .math import atan   #DEFINE_ALIAS
-from .math import logsumexp  #DEFINE_ALIAS
-# from .math import inverse   #DEFINE_ALIAS
-from .math import log1p  #DEFINE_ALIAS
-# from .math import erf   #DEFINE_ALIAS
-from .math import addcmul  #DEFINE_ALIAS
-from .math import addmm  #DEFINE_ALIAS
-from .math import clamp  #DEFINE_ALIAS
+
 # from .attribute import rank   #DEFINE_ALIAS
 # from .attribute import shape   #DEFINE_ALIAS
 # from .io import save   #DEFINE_ALIAS
@@ -164,15 +112,3 @@ from .manipulation import unsqueeze  #DEFINE_ALIAS
 from .manipulation import flip  #DEFINE_ALIAS
 # from .manipulation import unbind   #DEFINE_ALIAS
 from .manipulation import roll  #DEFINE_ALIAS
-from .search import argmax  #DEFINE_ALIAS
-# from .search import argmin   #DEFINE_ALIAS
-# from .search import argsort   #DEFINE_ALIAS
-# from .search import has_inf   #DEFINE_ALIAS
-# from .search import has_nan   #DEFINE_ALIAS
-# from .search import masked_select   #DEFINE_ALIAS
-# from .search import topk   #DEFINE_ALIAS
-from .search import where  #DEFINE_ALIAS
-from .search import index_select  #DEFINE_ALIAS
-from .search import index_sample  # DEFINE_ALIAS
-from .search import nonzero  #DEFINE_ALIAS
-from .search import sort  #DEFINE_ALIAS
